requirements.txt
streamlit==1.29.0
pandas
plotly-express
